+++ lib/core/methods/json_dump.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
-# Copyright (C) 2015 ToolsWatch.org
-# This file is part of vFeed Vulnerability Database Community API Parser - http://www.toolswatch.org
+# Copyright (C) 2016 ToolsWatch.org
+# This file is part of vFeed Correlated Threat & Vulnerability Community Database API Parser - http://www.toolswatch.org
 # See the file 'LICENSE' for copying permission.
 
 import os