+++ vfeedcli.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-# Copyright (C) 2015 ToolsWatch.org
+# Copyright (C) 2016 ToolsWatch.org
 # This file is part of vFeed Vulnerability Database Community API Parser - http://www.toolswatch.org
 # See the file 'LICENSE' for copying permission.
 
@@ -11,6 +11,7 @@ try:
     from lib.common.banner import banner
     from lib.core.search import Search
     from lib.core.update import Update
+    from lib.migration.mongo import Migrate
     from config.constants import build, title
     from lib.common.utils import enum_functions, enum_classes
 except ImportError, e:
@@ -29,9 +30,11 @@ if __name__ == "__main__":
                         help="Search utility for CVE,CPE,CWE, OVAL or free text")
     parser.add_argument("--stats", metavar="get_stats / get_latest", type=str,
                         help="View the vFeed Database statistics", nargs=1)
-    parser.add_argument("-u", "--update", help="Update the Vulnerability and Threat Database", action="store_true", required=False)
+    parser.add_argument("-u", "--update", help="Update the Vulnerability and Threat Database", action="store_true",
+                        required=False)
     parser.add_argument("--list", help="Enumerate the list of available methods", action="store_true", required=False)
     parser.add_argument("--banner", help="Print vFeed banner", action="store_true", required=False)
+    parser.add_argument("--migrate", help="Migration to MongoDB", action="store_true", required=False)
     args = parser.parse_args()
 
     if args.search:
@@ -40,6 +43,8 @@ if __name__ == "__main__":
         Update().update()
     elif args.banner:
         banner()
+    elif args.migrate:
+        Migrate()
     elif args.stats:
         method_name = args.stats[0]
         if method_name == "get_stats" or method_name == "get_latest":
@@ -62,5 +67,3 @@ if __name__ == "__main__":
                 enum_classes(method_name, cve_id)
             else:
                 print"[!] Unknown Method"
-
-